icar_web/icar_back/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.1
flask-cors==6.0.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.0.0
python-qpid-proton==0.37.0
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
